Build Grammar object from parsed PEG file
Introduce the ability to construct a new Grammar object from the results
of parsing a grammar file in the PEG syntax. This allows us to write
custom grammars to support arbitrary text-based file formats.

The grammar building process works by traversing a tree of Term objects
produced as a result of a parse. As each instance of the Definition
production is constructed, an Expression object is created for its body
based corresponding portion of the parse tree, and a new definition is
added to the Grammar object being built.

The resulting grammar can then be used to parse a file in the syntax
defined by that grammar - though this is not yet supported as a
command-line option.

Currently, a large number of assert statements are used when processing
the Term tree. These sanity checks ensure that the terms match what we
expect from the grammar. It is anticipated that in the future it will be
possible to write processing code in a more type-safe manner, both to a
limited extent in C, and using a dedicated tree-transformation language
to be integrated with the parsing process. The current parser is
essentially an interpreter which we can extend with additional language
constructs to support this transformation.