Add support for String expressions
Parsing expression grammars (PEGs) make no distinction between lexical
and syntactic analysis. The same constructs are used for expressing the
what "tokens" (e.g. identifier or symbol) and "productions" (e.g.
expression) look like.

Because of the simplistic parse tree construction method we've used to
date, the tree resulting from a parse contains a very large number of
nodes, many of which represent components of what would normally be
considered "tokens". These include leaf nodes for every single character
found, and interior nnodes for things like choice and repetition
constructs.

All these extra nodes are generally not of interest - it is usually more
convenient to have, for example, an Identifier as a single node in the
tree, represented by its string value.

With this commit, the grammar format now supports expressions of the
form $(...), where the inner expression will result in a single node
being generated in the parse tree. This does not affect the set of
strings accepted for a given language, but is merely an instruction to
the parser on how to represent what it finds.