More flexible command-line options
Use a more generic model of command-line options which moves away from
specific use cases.  As before, the command can be run with two
filenames, GRAMMAR and INPUT, to parse a file.

The -b (or --builtin) option is now used to select the builtin grammar
instead of a user-specified one. In this case, the GRAMMAR file is not
needed.

The -g (or --grammar) option is now used to ignore the input and instead
show the grammar that would have been used to parse it. In this case,
the INPUT file is not needed.