Option to print expressions as trees
When printing out a grammar definition, add the ability to show the body
of each rule as an expression tree, with one node per line. This makes
it easier to understand the structure of a grammar. In the future, it
will also be useful for displaying information associated with each
node, such as when we implement type inference.

As an example, the following rule in the existing concrete format:

    Term    : $Add(Factor PLUS Term)
            | $Sub(Factor MINUS Term)
            | Factor;

look as follows when printed with the -e (or --exprtree) option:

    Term

        Choice
        |-- Label "Add"
        |   \-- Sequence
        |       |-- Ident "Factor"
        |       |-- Ident "PLUS"
        |       \-- Ident "Term"
        |-- Label "Sub"
        |   \-- Sequence
        |       |-- Ident "Factor"
        |       |-- Ident "MINUS"
        |       \-- Ident "Term"
        \-- Ident "Factor"